Data to validate the ALAMO flowsheet
This data consists on full space instances that did not converge to optimality. The goal is to show that the results given by the ALAMO flowsheet exist by proving that the square system solves in these cases. The inputs to this square system are the degrees of freedom determined by the ALAMO flowsheet.