Data to validate the ALAMO Flowsheet
Data consists on full space instances (T,P) that did not converge to optimality. The goal is to show that the results given by the ALAMO Flowsheet exist by proving that the square system solves in these cases. the inputs to the square system are the DOF determined by the ALAMO Flowsheet.